+++ PyHarm.py
@@ -1,3 +1,5 @@
+from Data_funkce import *
+
 class Akord:
     def __init__(self, chord):
         if len(chord) > 1:
@@ -77,50 +79,11 @@ class Scale:
             out.append(self.tony[i]+ends[i])
         return out
 
-def varuj(text):
-    print('\x1b[0;31;40m' + "[WARNING]" + '\x1b[0m' + text)
-
-
-def to_num(ton):
-    out = 0
-    tony = ["C", "C#", "Db", "D", "D#", "Eb", "E", "F", "F#", "Gb", "G", "G#", "Ab", "A", "A#", "Bb", "Hb", "B", "H"]
-    valu = [0,   1,    1,    2,   3,    3,    4,   5,   6,    6,    7,   8,    8,    9,    10,   10,  10,   11,  11 ]
-    for i in range(len(tony)):
-        if ton == tony[i]:
-            out = valu[i]
-    return out
-
-def to_note(ton):
-    tony = ["C", "C#", "D", "Eb", "E", "F", "F#", "G", "Ab", "A", "Bb", "B"]
-    return tony[ton]
-
-def to_roman(num, typ):
-    if num == 1:
-        roman = "i"
-    elif num == 2:
-        roman = "ii"
-    elif num == 3:
-        roman = "iii"
-    elif num == 4:
-        roman = "iv"
-    elif num == 5:
-        roman = "v"
-    elif num == 6:
-        roman = "vi"
-    elif num == 7:
-        roman = "vii"
-    if typ == "Major":
-        roman = roman.upper()
-    if typ == "dim":
-        roman += "°"
-    if typ == "Dom7":
-        roman = roman.upper() + "7"
-    return roman
-
 
 def analyse(para):
     akordy = []
-    out = ""
+    out = []
+    ans = ""
     for i in para:
         akordy.append(Akord(i))
     first = akordy[0]
@@ -128,10 +91,15 @@ def analyse(para):
     legal = stupnice.akordy()
     for akord in akordy:
         if not akord.chord() in legal:
-            out += akord.chord() + "? "
+            out.append("OFF_KEY")
         else:
-            out += to_roman((legal.index(akord.chord()) + 1), akord.quality) + " "
-    print(out)
+            out.append(to_roman((legal.index(akord.chord()) + 1), akord.quality))
+    for i in range(len(out)):
+        if out[i] == "OFF_KEY":
+            pass
+    for i in out:
+        ans += i + " "
+    print(ans)
 
 
 def execute(inp):
@@ -148,21 +116,21 @@ def execute(inp):
         print("\nPříkaz " + com + " neexistuje!\nPro seznam příkazů použijte příkaz help")
     return 0
 
+if __name__ == "__main__":
+    print("PyHarm ver. 1.0\n"
+        "Analyzátor a generátor akordových postupů\n"
+        "Pro seznam příkazů použijte příkaz help\n"
+        "By VP 2021\n")
 
-print("PyHarm ver. 1.0\n"
-      "Analyzátor a generátor akordových postupů\n"
-      "Pro seznam příkazů použijte příkaz help\n"
-      "By VP 2021\n")
-
-run = True
-while run:
-    try:
-        inp = input("PyHarm>")
-        ret = execute(inp)
-        if ret == "term":
-            run = False
-    except KeyboardInterrupt:
-        print("Pro ukončení programu je třeba použít příkaz quit")
-    except:
-        varuj("ERROR 1")
-print("Program ukončen")
+    run = True
+    while run:
+        try:
+            inp = input("PyHarm>")
+            ret = execute(inp)
+            if ret == "term":
+                run = False
+        except KeyboardInterrupt:
+            print("Pro ukončení programu je třeba použít příkaz quit")
+        except:
+            varuj("ERROR 1")
+    print("Program ukončen")
